Finished hw 4

+++ services/contactService.js
@@ -1,14 +1,14 @@
 const { Contact } = require('../db/contactModel');
 const {NotFoundError} = require('../helpers/error')
 
-const getContacts = async () => {
-    const contacts = await Contact.find({})
+const getContacts = async (userId) => {
+    const contacts = await Contact.find({owner : userId})
 
     return contacts
 }
 
-const getContactById = async (contactId) => {
-    const contact = await Contact.findById(contactId)
+const getContactById = async (contactId , userId) => {
+    const contact = await Contact.findOne({_id: contactId, owner : userId})
 
     if (!contact) {
         throw new NotFoundError('Not found contact')
@@ -17,7 +17,7 @@ const getContactById = async (contactId) => {
     return contact
 }
 
-const addContact = async (body) => {
+const addContact = async (body , userId) => {
 const {
     name,
     email,
@@ -25,16 +25,16 @@ const {
     favorite
     } = body;
 
-    const contact = new Contact({ name, email, phone, favorite })
+    const contact = new Contact({ name, email, phone, favorite , owner: userId})
     await contact.save()
 
     return contact
 }
 
-const updateContact = async (contactId, {name, email ,phone , favorite } ) => {
+const updateContact = async (contactId, {name, email ,phone , favorite } , userId) => {
 
    const updatedContact =  await Contact.findOneAndUpdate(
-      {_id: contactId },
+      {_id: contactId , owner: userId },
       {$set: {name, email ,phone , favorite }}
     );
     
@@ -42,16 +42,16 @@ const updateContact = async (contactId, {name, email ,phone , favorite } ) => {
 
 }
 
-const deleteContact = async (contactId) => {
+const deleteContact = async (contactId , userId) => {
 
-  await Contact.findOneAndRemove({_id: contactId}); 
+  await Contact.findOneAndRemove({_id: contactId ,  owner: userId}); 
     
 }
 
-const updateStatusContact = async (contactId , {favorite}) => {
+const updateStatusContact = async (contactId , {favorite} , userId) => {
 
  const contact = await Contact.findOneAndUpdate(
-      {_id: contactId },
+      {_id: contactId ,  owner: userId},
       {$set: { favorite }}
     );
